Split out MessageChannel
